% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgy.R
\name{confint.phylandml}
\alias{confint.phylandml}
\title{Profile confidence intervals for fitted phylogeographic model}
\usage{
confint.phylandml(fit, whichparm, guess_se, np = 13, ncpu = 1, ...)
}
\arguments{
\item{fit}{A fited model of class *phylandml*}

\item{whichparm}{Name of parameter to be profiled}

\item{guess_se}{Initial guess of standard error of parameter}

\item{np}{Integer number of spline points to be used for interpolation. Increase this value to get more accurate profile}

\item{ncpu}{Number of CPUs to use on multi-core machines}
}
\value{
Likelihood profile
}
\description{
Profile confidence intervals for fitted phylogeographic model
}
