\name{phyland-package}
\alias{phyland-package}
\alias{phyland}
\docType{package}
\title{
\packageTitle{phyland}
}
\description{
\packageDescription{phyland}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{phyland}
\packageIndices{phyland}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{phyland}

Maintainer: \packageMaintainer{phyland}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
