% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgy2.R
\name{phylandml}
\alias{phylandml}
\title{Maximum likelihood inference of migration rates and effective population size between demes}
\usage{
phylandml(tree, delimiter = "_", index = NULL, regex = NULL,
  design = NULL, method = "BFGS", quiet = FALSE, ...)
}
\arguments{
\item{tree}{A time-scaled phylogeny in ape::phylo format}

\item{delimiter}{A character string which will be used to infer the deme of sampling from tip labels}

\item{index}{The integer position within tip labels of the sample deme information. If not provided will use the last position, e.g. <accession number>_locationOfSampling}

\item{regex}{As an alternative to delimiter/index, the location of sampling can be inferred by matching a regular expression string}

\item{design}{A matrix describing which migration rates to estimate and which migration rates are assumed to be equal. Row and column names should correspond to names of demes. Unique integers should specify rates to estimate. If not provided, all rates will be estimated
method Optimisation method to be used by optim
quiet If TRUE, will suppress likelihood printing to stdout}

\item{...}{Additional parameters passed to optim}
}
\value{
A fitted model
}
\description{
Maximum likelihood inference of migration rates and effective population size between demes
}
