% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgy.R
\name{confint.phylandml}
\alias{confint.phylandml}
\title{Profile confidence intervals for fitted phylogeographic model}
\usage{
confint.phylandml(fit, whichparms, ncpu = 1, tol.newmin = 0.1, ...)
}
\arguments{
\item{fit}{A fited model of class *phylandml*}

\item{whichparms}{A vector of type character naming parameters to be profiled. A single parameter is allowed}

\item{ncpu}{Will spread the profiling job across multiple processing units of ncpu > 1}

\item{tol.newmin}{Tolerance in log likelihood units for detecting better optima}

\item{...}{Additional parameters passed to bbmle::confint.mle2}
}
\value{
A fitted model
}
\description{
Profile confidence intervals for fitted phylogeographic model
}
